requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
langchain-google-genai
langchain-core
llama-index
google-generativeai
loguru
pytest
celery
redis
